Pdf_Report_Generator/requirements.txt
pandas
folium
matplotlib
numpy
openpyxl
shapely
playwright
groq
reportlab
python-dotenv
pypdf
fastapi 
uvicorn

affine==2.4.0
alembic==1.17.1
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.17.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.8.2
bottle==0.13.2
bottle-websocket==0.2.9
branca==0.8.1
cachetools==5.5.0
catboost==1.2.8
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
choreographer==1.1.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.2.2
contextily==1.6.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dash==3.0.0
dash-bootstrap-components==2.0.0
dash-leaflet==1.0.15
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.8.0
Eel==0.18.1
eli5==0.16.0
email-validator==2.3.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.16
fastapi-cloud-cli==0.3.1
fastjsonschema==2.21.1
fiona==1.10.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
folium==0.19.4
fonttools==4.55.3
fqdn==1.5.1
future==1.0.0
fuzzywuzzy==0.18.0
GeoAlchemy2==0.18.0
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
geovoronoi
gevent==24.11.1
gevent-websocket==0.10.1
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
gmplot==1.4.1
googlemaps==4.10.0
graphviz==0.20.3
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.7.1
httpx==0.28.1
idna==3.10
imbalanced-learn==0.14.0
imblearn==0.0
importlib_metadata==8.6.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.17.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaleido==1.1.0
kiwisolver==1.4.8
Levenshtein==0.27.1
lightgbm==4.6.0
logistro==1.1.0
lxml==5.3.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mercantile==1.2.1
mistune==3.1.2
mpld3==0.5.10
mysql-connector==2.2.9
mysqlclient==2.2.7
narwhals==1.22.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.3
notebook_shim==0.2.4
numpy==2.2.1
openpyxl==3.1.5
orjson==3.11.3
osmnx==2.0.6
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pefile==2023.2.7
pillow==11.1.0
platformdirs==4.3.6
plotly==6.3.0
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyMySQL==1.1.2
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1 
pyqtgraph==0.13.7
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
pytest==8.3.4
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-Levenshtein==0.27.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.3.0
RapidFuzz==3.13.0
rasterio==1.4.3
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.15.1
rignore==0.7.6
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.44.0
setuptools==79.0.1
shapely==2.0.6
shellingham==1.5.4
simplejson==3.20.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.45.3
statsmodels==0.14.5
stop-words==2018.7.23
streamlit==1.41.1
streamlit-autorefresh==1.0.1
streamlit_folium==0.24.0
stringcase==1.2.0
sv-ttk==2.6.0
tabulate==0.9.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tkcalendar==1.6.1
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.11.0
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.6
wheel==0.45.1
widgetsnbextension==4.0.13
xgboost==3.0.0
XlsxWriter==3.2.0
xyzservices==2025.1.0
yellowbrick==1.5
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
